Объединил два приложение в один репозиторий
