+++ backend/bake_cake/bot/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class BotConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'bot'
